refactor: move openAI gym depending code into a submodule
