+++ pyrlutils/transition.py
@@ -3,7 +3,6 @@ from types import LambdaType
 from typing import Tuple, Dict
 
 import numpy as np
-import gym
 
 from .state import DiscreteState, DiscreteStateValueType
 from .reward import IndividualRewardFunction
@@ -145,21 +144,3 @@ class TransitionProbabilityFactory:
     @property
     def objects_generated(self) -> bool:
         return self._objects_generated
-
-
-class OpenAIGymDiscreteEnvironmentTransitionProbabilityFactory(TransitionProbabilityFactory):
-    def __init__(self, envname):
-        super().__init__()
-        self.gymenv = gym.make(envname)
-        self._convert_openai_gymenv_to_transprob()
-
-    def _convert_openai_gymenv_to_transprob(self):
-        P = self.gymenv.env.P
-        for state_value, trans_dict in P.items():
-            new_trans_dict = {}
-            for action_value, next_state_list in trans_dict.items():
-                new_trans_dict[action_value] = [
-                    NextStateTuple(next_state[1], next_state[0], next_state[2], next_state[3])
-                    for next_state in next_state_list
-                ]
-            self.add_state_transitions(state_value, new_trans_dict)
